pyproject.toml
[project]
name = "madurez-digital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "loguru>=0.7.3",
  "msal>=1.33.0",
  "openpyxl>=3.1.5",
  "pandas>=2.3.2",
  "python-dotenv>=1.1.1",
  "streamlit>=1.49.1",
  "unidecode>=1.4.0",
]
